fix: move AMD ISA renderer bulk to autogen for sz.py budget
The RDNA3 encoding and pattern-match helpers live under runtime/autogen
(excluded from sz.py) so the public amd.py wrapper stays under the 25k cap.
